Split AllocateDiskSpaceModel from DiskStorageModel
Most of DiskStorageModel was specific to the Allocate disk space page.
Make it a pure service and split the page-specific parts as a view
model.
